LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-navigation-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.


Harness Error (TIMEOUT), message = null

TIMEOUT Navigate a window with javascript:-urls in enforcing mode. Test timed out
NOTRUN Navigate a window with javascript:-urls w/ default policy in enforcing mode.
NOTRUN Navigate a window with javascript:-urls in report-only mode.
NOTRUN Navigate a window with javascript:-urls w/ default policy in report-only mode.
NOTRUN Navigate a frame with javascript:-urls in enforcing mode.
NOTRUN Navigate a frame with javascript:-urls w/ default policy in enforcing mode.
NOTRUN Navigate a frame with javascript:-urls in report-only mode.
NOTRUN Navigate a frame with javascript:-urls w/ default policy in report-only mode.
